Stop duplicating provider-turn runtime flow inside the coordinator
The coordinator still carried a second copy of the provider-turn runtime shell even after the apply/finalization extraction. This change makes provider_turn_runtime.rs the authoritative home for provider-turn request handling, continuation state, checkpoint shaping, observer/runtime glue, and tool-intent scoping so turn_coordinator.rs can keep collapsing toward orchestration only.

Constraint: Preserve the existing provider-turn followup loop and checkpoint semantics while reducing duplication
Rejected: Jump straight to safe-lane or chat-shell cleanup first | the provider-turn runtime duplication was the more direct blocker to a small loop kernel
Confidence: high
Scope-risk: moderate
Reversibility: clean
Directive: Keep provider-turn runtime helpers and state types in provider_turn_runtime.rs; do not reintroduce shadow copies inside turn_coordinator.rs
Tested: ./scripts/cargo-local-toolchain.sh check -p loong-app; ./scripts/cargo-local-toolchain.sh test -p loong-app scope_provider_turn_tool_intents -- --test-threads=1; ./scripts/cargo-local-toolchain.sh test -p loong-app build_provider_turn_tool_terminal_events -- --test-threads=1; ./scripts/cargo-local-toolchain.sh test -p loong-app provider_turn_continue_phase_checkpoint -- --test-threads=1; ./scripts/cargo-local-toolchain.sh fmt --all -- --check; ./scripts/cargo-local-toolchain.sh clippy -p loong-app --all-targets -- -D warnings
Not-tested: full workspace test suite; daemon/control-plane surfaces in this commit (unchanged by this slice)
