+++ crates/app/src/conversation/turn_coordinator.rs
@@ -21,6 +21,8 @@ mod pending_approval;
 mod provider_turn_apply;
 #[path = "turn_coordinator/provider_turn_reply.rs"]
 mod provider_turn_reply;
+#[path = "turn_coordinator/provider_turn_runtime.rs"]
+mod provider_turn_runtime;
 #[path = "turn_coordinator/safe_lane_events.rs"]
 mod safe_lane_events;
 #[path = "turn_coordinator/safe_lane_execution.rs"]
@@ -56,6 +58,7 @@ use crate::task_progress::{
 use self::pending_approval::*;
 use self::provider_turn_apply::*;
 use self::provider_turn_reply::*;
+use self::provider_turn_runtime::*;
 use self::safe_lane_events::*;
 use self::safe_lane_execution::*;
 use self::safe_lane_governor::*;
@@ -625,361 +628,10 @@ fn provider_turn_tool_name_signature(intents: &[ToolIntent]) -> String {
         .join("||")
 }
 
-#[derive(Debug, Clone)]
-struct ProviderTurnContinuePhase {
-    request: TurnCheckpointRequest,
-    lane_execution: ProviderTurnLaneExecution,
-    reply_phase: ToolDrivenReplyPhase,
-    loop_verdict: Option<ProviderTurnLoopVerdict>,
-    followup_config: LoongConfig,
-    ingress: Option<ConversationIngressContext>,
-}
-
-impl ProviderTurnContinuePhase {
-    fn new(
-        tool_intents: usize,
-        lane_execution: ProviderTurnLaneExecution,
-        loop_verdict: Option<ProviderTurnLoopVerdict>,
-        followup_config: LoongConfig,
-        ingress: Option<&ConversationIngressContext>,
-    ) -> Self {
-        let reply_phase = lane_execution.reply_phase();
-        Self {
-            request: TurnCheckpointRequest::Continue { tool_intents },
-            lane_execution,
-            reply_phase,
-            loop_verdict,
-            followup_config,
-            ingress: ingress.cloned(),
-        }
-    }
-
-    fn checkpoint(
-        &self,
-        preparation: &ProviderTurnPreparation,
-        user_input: &str,
-        reply: &str,
-    ) -> TurnCheckpointSnapshot {
-        self.checkpoint_with_continuation_state(preparation, user_input, reply, None)
-    }
-
-    fn checkpoint_with_continuation_state(
-        &self,
-        preparation: &ProviderTurnPreparation,
-        user_input: &str,
-        reply: &str,
-        continuation_state: Option<ToolDrivenContinuationState>,
-    ) -> TurnCheckpointSnapshot {
-        let reply_checkpoint = if continuation_state.is_some() {
-            TurnReplyCheckpoint::from_phase_with_continuation_state(
-                &self.reply_phase,
-                continuation_state,
-            )
-        } else {
-            TurnReplyCheckpoint::from_phase(&self.reply_phase)
-        };
-        build_resolved_provider_checkpoint(
-            preparation,
-            user_input,
-            Some(reply),
-            self.request.clone(),
-            Some(self.lane_execution.checkpoint()),
-            Some(reply_checkpoint),
-            TurnFinalizationCheckpoint::persist_reply(ReplyPersistenceMode::Success),
-        )
-    }
-
-    fn tool_intent_count(&self) -> usize {
-        match self.request {
-            TurnCheckpointRequest::Continue { tool_intents } => tool_intents,
-            TurnCheckpointRequest::FinalizeInlineProviderError
-            | TurnCheckpointRequest::ReturnError => 0,
-        }
-    }
-
-    fn loop_warning_reason(&self) -> Option<&str> {
-        match self.loop_verdict.as_ref() {
-            Some(ProviderTurnLoopVerdict::InjectWarning { reason }) => Some(reason.as_str()),
-            _ => None,
-        }
-    }
-
-    fn hard_stop_reason(&self) -> Option<&str> {
-        match self.loop_verdict.as_ref() {
-            Some(ProviderTurnLoopVerdict::HardStop { reason }) => Some(reason.as_str()),
-            _ => None,
-        }
-    }
-
-    async fn resolve<R: ConversationRuntime + ?Sized>(
-        &self,
-        runtime: &R,
-        session_id: &str,
-        preparation: &ProviderTurnPreparation,
-        user_input: &str,
-        turn_loop_policy: &ProviderTurnLoopPolicy,
-        turn_loop_state: &mut ProviderTurnLoopState,
-        remaining_provider_rounds: usize,
-        binding: ConversationRuntimeBinding<'_>,
-        observer: Option<&ConversationTurnObserverHandle>,
-        retry_progress: crate::provider::ProviderRetryProgressCallback,
-    ) -> ResolvedProviderTurn {
-        resolve_provider_turn_reply(
-            runtime,
-            &self.followup_config,
-            session_id,
-            preparation,
-            self,
-            user_input,
-            turn_loop_policy,
-            turn_loop_state,
-            remaining_provider_rounds,
-            binding,
-            self.ingress.as_ref(),
-            observer,
-            retry_progress,
-        )
-        .await
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum ResolvedProviderTurn {
-    PersistReply(ResolvedProviderReply),
-    ReturnError(ResolvedProviderError),
-}
-
-impl ResolvedProviderTurn {
-    fn persist_reply(
-        reply: String,
-        usage: Option<Value>,
-        checkpoint: TurnCheckpointSnapshot,
-    ) -> Self {
-        Self::PersistReply(ResolvedProviderReply {
-            reply,
-            usage,
-            checkpoint,
-        })
-    }
-
-    fn return_error(error: String, checkpoint: TurnCheckpointSnapshot) -> Self {
-        Self::ReturnError(ResolvedProviderError { error, checkpoint })
-    }
-
-    #[cfg(test)]
-    fn checkpoint(&self) -> &TurnCheckpointSnapshot {
-        match self {
-            Self::PersistReply(reply) => &reply.checkpoint,
-            Self::ReturnError(error) => &error.checkpoint,
-        }
-    }
-
-    fn terminal_phase<'a>(
-        &'a self,
-        session: &ProviderTurnSessionState,
-    ) -> ProviderTurnTerminalPhase<'a> {
-        match self {
-            Self::PersistReply(reply) => {
-                ProviderTurnTerminalPhase::PersistReply(ProviderTurnPersistReplyPhase {
-                    checkpoint: &reply.checkpoint,
-                    tail_phase: ProviderTurnReplyTailPhase::from_session(
-                        session,
-                        reply.reply.as_str(),
-                    ),
-                    usage: reply.usage.clone(),
-                })
-            }
-            Self::ReturnError(error) => {
-                ProviderTurnTerminalPhase::ReturnError(ProviderTurnReturnErrorPhase {
-                    checkpoint: &error.checkpoint,
-                    error: error.error.as_str(),
-                })
-            }
-        }
-    }
-
-    #[cfg(test)]
-    fn reply_text(&self) -> Option<&str> {
-        match self {
-            Self::PersistReply(reply) => Some(reply.reply.as_str()),
-            Self::ReturnError(_) => None,
-        }
-    }
-
-    fn provider_error_text(&self) -> Option<&str> {
-        match self {
-            Self::PersistReply(reply) => provider_error_reply_body(reply.reply.as_str()),
-            Self::ReturnError(error) => Some(error.error.as_str()),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ResolvedProviderReply {
-    reply: String,
-    usage: Option<Value>,
-    checkpoint: TurnCheckpointSnapshot,
-}
-
 #[derive(Debug, Clone, PartialEq)]
 pub struct ConversationTurnOutcome {
     pub reply: String,
     pub usage: Option<Value>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ResolvedProviderError {
-    error: String,
-    checkpoint: TurnCheckpointSnapshot,
-}
-
-#[derive(Debug)]
-enum ProviderTurnTerminalPhase<'a> {
-    PersistReply(ProviderTurnPersistReplyPhase<'a>),
-    ReturnError(ProviderTurnReturnErrorPhase<'a>),
-}
-
-impl<'a> ProviderTurnTerminalPhase<'a> {
-    async fn apply<R: ConversationRuntime + ?Sized>(
-        self,
-        config: &LoongConfig,
-        runtime: &R,
-        session_id: &str,
-        user_input: &str,
-        binding: ConversationRuntimeBinding<'_>,
-    ) -> CliResult<ConversationTurnOutcome> {
-        match self {
-            Self::PersistReply(phase) => {
-                finalize_provider_turn_reply(
-                    config,
-                    runtime,
-                    session_id,
-                    user_input,
-                    &phase.tail_phase,
-                    phase.usage,
-                    phase.checkpoint,
-                    binding,
-                )
-                .await
-            }
-            Self::ReturnError(phase) => {
-                persist_resolved_provider_error_checkpoint(
-                    runtime,
-                    session_id,
-                    phase.checkpoint,
-                    binding,
-                )
-                .await?;
-                Err(phase.error.to_owned())
-            }
-        }
-    }
-}
-
-#[derive(Debug)]
-struct ProviderTurnPersistReplyPhase<'a> {
-    checkpoint: &'a TurnCheckpointSnapshot,
-    tail_phase: ProviderTurnReplyTailPhase,
-    usage: Option<Value>,
-}
-
-#[derive(Debug)]
-struct ProviderTurnReturnErrorPhase<'a> {
-    checkpoint: &'a TurnCheckpointSnapshot,
-    error: &'a str,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum ProviderTurnRequestTerminalPhase {
-    PersistInlineProviderError { reply: String },
-    ReturnError { error: String },
-}
-
-impl ProviderTurnRequestTerminalPhase {
-    fn persist_inline_provider_error(reply: String) -> Self {
-        Self::PersistInlineProviderError { reply }
-    }
-
-    fn return_error(error: String) -> Self {
-        Self::ReturnError { error }
-    }
-
-    fn resolve(
-        self,
-        preparation: &ProviderTurnPreparation,
-        user_input: &str,
-    ) -> ResolvedProviderTurn {
-        match self {
-            Self::PersistInlineProviderError { reply } => {
-                let checkpoint = build_resolved_provider_checkpoint(
-                    preparation,
-                    user_input,
-                    Some(reply.as_str()),
-                    TurnCheckpointRequest::FinalizeInlineProviderError,
-                    None,
-                    None,
-                    TurnFinalizationCheckpoint::persist_reply(
-                        ReplyPersistenceMode::InlineProviderError,
-                    ),
-                );
-                ResolvedProviderTurn::persist_reply(reply, None, checkpoint)
-            }
-            Self::ReturnError { error } => {
-                let checkpoint = build_resolved_provider_checkpoint(
-                    preparation,
-                    user_input,
-                    None,
-                    TurnCheckpointRequest::ReturnError,
-                    None,
-                    None,
-                    TurnFinalizationCheckpoint::ReturnError,
-                );
-                ResolvedProviderTurn::return_error(error, checkpoint)
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-struct SafeLaneTurnOutcome {
-    result: TurnResult,
-    terminal_route: Option<SafeLaneFailureRoute>,
-}
-
-impl SafeLaneTurnOutcome {
-    fn without_terminal_route(result: TurnResult) -> Self {
-        Self {
-            result,
-            terminal_route: None,
-        }
-    }
-
-    fn with_terminal_route(result: TurnResult, terminal_route: SafeLaneFailureRoute) -> Self {
-        Self {
-            result,
-            terminal_route: Some(terminal_route),
-        }
-    }
-}
-
-fn build_resolved_provider_checkpoint(
-    preparation: &ProviderTurnPreparation,
-    user_input: &str,
-    reply_text: Option<&str>,
-    request: TurnCheckpointRequest,
-    lane: Option<TurnLaneExecutionSnapshot>,
-    reply: Option<TurnReplyCheckpoint>,
-    finalization: TurnFinalizationCheckpoint,
-) -> TurnCheckpointSnapshot {
-    TurnCheckpointSnapshot {
-        identity: reply_text
-            .map(|assistant_reply| TurnCheckpointIdentity::from_turn(user_input, assistant_reply)),
-        preparation: preparation.checkpoint(),
-        request,
-        lane,
-        reply,
-        finalization,
-    }
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -2581,419 +2233,6 @@ fn observe_turn_phase(
     };
 
     observer.on_phase(event);
-}
-
-fn observe_non_provider_turn_terminal_success_phases(
-    observer: Option<&ConversationTurnObserverHandle>,
-) {
-    let finalizing_event = ConversationTurnPhaseEvent {
-        phase: ConversationTurnPhase::FinalizingReply,
-        provider_round: None,
-        lane: None,
-        tool_call_count: 0,
-        message_count: None,
-        estimated_tokens: None,
-    };
-    observe_turn_phase(observer, finalizing_event);
-
-    let completed_event = ConversationTurnPhaseEvent {
-        phase: ConversationTurnPhase::Completed,
-        provider_round: None,
-        lane: None,
-        tool_call_count: 0,
-        message_count: None,
-        estimated_tokens: None,
-    };
-    observe_turn_phase(observer, completed_event);
-}
-
-fn observe_provider_turn_tool_batch_started(
-    observer: Option<&ConversationTurnObserverHandle>,
-    turn: &ProviderTurn,
-) {
-    let Some(observer) = observer else {
-        return;
-    };
-
-    for intent in &turn.tool_intents {
-        let tool_name = effective_result_tool_name(intent);
-        let request_summary = summarize_single_tool_followup_request(intent);
-        let event = ConversationTurnToolEvent::running(intent.tool_call_id.clone(), tool_name)
-            .with_request_summary(request_summary);
-        observer.on_tool(event);
-    }
-}
-
-fn observe_provider_turn_tool_batch_terminal(
-    observer: Option<&ConversationTurnObserverHandle>,
-    tool_events: &[ConversationTurnToolEvent],
-) {
-    let Some(observer) = observer else {
-        return;
-    };
-
-    for tool_event in tool_events {
-        observer.on_tool(tool_event.clone());
-    }
-}
-
-fn build_provider_turn_tool_terminal_events(
-    turn: &ProviderTurn,
-    turn_result: &TurnResult,
-    trace: Option<&ToolBatchExecutionTrace>,
-) -> Vec<ConversationTurnToolEvent> {
-    let mut trace_events = BTreeMap::new();
-    if let Some(trace) = trace {
-        for intent_outcome in &trace.intent_outcomes {
-            let event = match intent_outcome.status {
-                ToolBatchExecutionIntentStatus::Completed => ConversationTurnToolEvent::completed(
-                    intent_outcome.tool_call_id.clone(),
-                    intent_outcome.tool_name.clone(),
-                    intent_outcome.detail.clone(),
-                ),
-                ToolBatchExecutionIntentStatus::NeedsApproval => {
-                    let detail = intent_outcome.detail.clone().unwrap_or_default();
-                    ConversationTurnToolEvent::needs_approval(
-                        intent_outcome.tool_call_id.clone(),
-                        intent_outcome.tool_name.clone(),
-                        detail,
-                    )
-                }
-                ToolBatchExecutionIntentStatus::Denied => {
-                    let detail = intent_outcome.detail.clone().unwrap_or_default();
-                    ConversationTurnToolEvent::denied(
-                        intent_outcome.tool_call_id.clone(),
-                        intent_outcome.tool_name.clone(),
-                        detail,
-                    )
-                }
-                ToolBatchExecutionIntentStatus::Failed => {
-                    let detail = intent_outcome.detail.clone().unwrap_or_default();
-                    ConversationTurnToolEvent::failed(
-                        intent_outcome.tool_call_id.clone(),
-                        intent_outcome.tool_name.clone(),
-                        detail,
-                    )
-                }
-            };
-            trace_events.insert(intent_outcome.tool_call_id.clone(), event);
-        }
-    }
-
-    let mut events = Vec::new();
-    let mut unresolved_failure_emitted = false;
-
-    for intent in &turn.tool_intents {
-        if let Some(event) = trace_events.remove(intent.tool_call_id.as_str()) {
-            let request_summary = summarize_single_tool_followup_request(intent);
-            let event = event.with_request_summary(request_summary);
-            events.push(event);
-            continue;
-        }
-
-        let tool_name = effective_result_tool_name(intent);
-        let fallback_event = match turn_result {
-            TurnResult::FinalText(_)
-            | TurnResult::StreamingText(_)
-            | TurnResult::StreamingDone(_) => Some(ConversationTurnToolEvent::completed(
-                intent.tool_call_id.clone(),
-                tool_name,
-                None,
-            )),
-            TurnResult::NeedsApproval(requirement) => {
-                if unresolved_failure_emitted {
-                    None
-                } else {
-                    unresolved_failure_emitted = true;
-                    Some(ConversationTurnToolEvent::needs_approval(
-                        intent.tool_call_id.clone(),
-                        tool_name,
-                        requirement.reason.clone(),
-                    ))
-                }
-            }
-            TurnResult::ToolDenied(failure) => {
-                if unresolved_failure_emitted {
-                    None
-                } else {
-                    unresolved_failure_emitted = true;
-                    Some(ConversationTurnToolEvent::denied(
-                        intent.tool_call_id.clone(),
-                        tool_name,
-                        failure.reason.clone(),
-                    ))
-                }
-            }
-            TurnResult::ToolError(failure) => {
-                if unresolved_failure_emitted {
-                    None
-                } else {
-                    unresolved_failure_emitted = true;
-                    Some(ConversationTurnToolEvent::failed(
-                        intent.tool_call_id.clone(),
-                        tool_name,
-                        failure.reason.clone(),
-                    ))
-                }
-            }
-            TurnResult::ProviderError(failure) => {
-                if unresolved_failure_emitted {
-                    None
-                } else {
-                    unresolved_failure_emitted = true;
-                    Some(ConversationTurnToolEvent::interrupted(
-                        intent.tool_call_id.clone(),
-                        tool_name,
-                        failure.reason.clone(),
-                    ))
-                }
-            }
-        };
-
-        if let Some(fallback_event) = fallback_event {
-            let request_summary = summarize_single_tool_followup_request(intent);
-            let fallback_event = fallback_event.with_request_summary(request_summary);
-            events.push(fallback_event);
-        }
-    }
-
-    events
-}
-
-#[cfg(test)]
-fn summarize_tool_event_request(intent: &ToolIntent) -> Option<String> {
-    summarize_single_tool_followup_request(intent)
-}
-fn provider_turn_observer_supports_streaming(
-    config: &LoongConfig,
-    observer: Option<&ConversationTurnObserverHandle>,
-) -> bool {
-    if observer.is_none() {
-        return false;
-    }
-
-    crate::provider::supports_turn_streaming_events(config)
-}
-
-async fn request_provider_turn_with_observer<R: ConversationRuntime + ?Sized>(
-    config: &LoongConfig,
-    runtime: &R,
-    session_id: &str,
-    turn_id: &str,
-    messages: &[Value],
-    tool_view: &crate::tools::ToolView,
-    binding: ConversationRuntimeBinding<'_>,
-    observer: Option<&ConversationTurnObserverHandle>,
-    retry_progress: crate::provider::ProviderRetryProgressCallback,
-) -> CliResult<ProviderTurn> {
-    if let Some(observer) = observer
-        && provider_turn_observer_supports_streaming(config, Some(observer))
-    {
-        let request_started_at = std::time::Instant::now();
-        let on_token = build_observer_streaming_token_callback(observer, request_started_at);
-        return runtime
-            .request_turn_streaming_with_retry_progress(
-                config,
-                session_id,
-                turn_id,
-                messages,
-                tool_view,
-                binding,
-                on_token,
-                retry_progress,
-            )
-            .await;
-    }
-
-    runtime
-        .request_turn_with_retry_progress(
-            config,
-            session_id,
-            turn_id,
-            messages,
-            tool_view,
-            binding,
-            retry_progress,
-        )
-        .await
-}
-
-async fn resolve_provider_turn<R: ConversationRuntime + ?Sized>(
-    config: &LoongConfig,
-    runtime: &R,
-    session_id: &str,
-    user_input: &str,
-    preparation: &ProviderTurnPreparation,
-    result: CliResult<ProviderTurn>,
-    error_mode: ProviderErrorMode,
-    binding: ConversationRuntimeBinding<'_>,
-    ingress: Option<&ConversationIngressContext>,
-    observer: Option<&ConversationTurnObserverHandle>,
-    retry_progress: crate::provider::ProviderRetryProgressCallback,
-) -> ResolvedProviderTurn {
-    let turn_loop_policy = ProviderTurnLoopPolicy::from_config(config);
-    let mut turn_loop_state = ProviderTurnLoopState::default();
-
-    match decide_provider_turn_request_action(result, error_mode) {
-        ProviderTurnRequestAction::Continue { turn } => {
-            let turn =
-                scope_provider_turn_tool_intents(turn, session_id, preparation.turn_id.as_str());
-            if let Some(reply) =
-                turn_loop_state.circuit_breaker_reply(&turn_loop_policy, turn.tool_intents.len())
-            {
-                return build_turn_loop_circuit_breaker_resolved_turn(
-                    preparation,
-                    user_input,
-                    turn.tool_intents.len(),
-                    reply,
-                );
-            }
-            let continue_phase = prepare_provider_turn_continue_phase(
-                config,
-                runtime,
-                session_id,
-                preparation,
-                turn,
-                &turn_loop_policy,
-                &mut turn_loop_state,
-                binding,
-                ingress,
-                observer,
-                1,
-                false,
-            )
-            .await;
-            continue_phase
-                .resolve(
-                    runtime,
-                    session_id,
-                    preparation,
-                    user_input,
-                    &turn_loop_policy,
-                    &mut turn_loop_state,
-                    config
-                        .conversation
-                        .turn_loop
-                        .max_discovery_followup_rounds
-                        .saturating_add(1)
-                        .max(1),
-                    binding,
-                    observer,
-                    retry_progress,
-                )
-                .await
-        }
-        ProviderTurnRequestAction::FinalizeInlineProviderError { reply } => {
-            ProviderTurnRequestTerminalPhase::persist_inline_provider_error(reply)
-                .resolve(preparation, user_input)
-        }
-        ProviderTurnRequestAction::ReturnError { error } => {
-            ProviderTurnRequestTerminalPhase::return_error(error).resolve(preparation, user_input)
-        }
-    }
-}
-
-fn scope_provider_turn_tool_intents(
-    mut turn: ProviderTurn,
-    session_id: &str,
-    turn_id: &str,
-) -> ProviderTurn {
-    for intent in &mut turn.tool_intents {
-        if intent.source.starts_with("provider_") {
-            // Provider-originated intents: runtime scope is authoritative.
-            intent.session_id = session_id.to_owned();
-            intent.turn_id = turn_id.to_owned();
-        } else {
-            // Non-provider intents: only fill in if missing.
-            if intent.session_id.trim().is_empty() {
-                intent.session_id = session_id.to_owned();
-            }
-            if intent.turn_id.trim().is_empty() {
-                intent.turn_id = turn_id.to_owned();
-            }
-        }
-    }
-    turn
-}
-
-fn provider_turn_usage(turn: &ProviderTurn) -> Option<Value> {
-    turn.raw_meta.get("usage").cloned()
-}
-
-fn build_turn_loop_circuit_breaker_resolved_turn(
-    preparation: &ProviderTurnPreparation,
-    user_input: &str,
-    tool_intents: usize,
-    reply: String,
-) -> ResolvedProviderTurn {
-    let checkpoint = build_resolved_provider_checkpoint(
-        preparation,
-        user_input,
-        Some(reply.as_str()),
-        TurnCheckpointRequest::Continue { tool_intents },
-        None,
-        None,
-        TurnFinalizationCheckpoint::persist_reply(ReplyPersistenceMode::Success),
-    );
-    ResolvedProviderTurn::persist_reply(reply, None, checkpoint)
-}
-
-async fn prepare_provider_turn_continue_phase<R: ConversationRuntime + ?Sized>(
-    config: &LoongConfig,
-    runtime: &R,
-    session_id: &str,
-    preparation: &ProviderTurnPreparation,
-    turn: ProviderTurn,
-    turn_loop_policy: &ProviderTurnLoopPolicy,
-    turn_loop_state: &mut ProviderTurnLoopState,
-    binding: ConversationRuntimeBinding<'_>,
-    ingress: Option<&ConversationIngressContext>,
-    observer: Option<&ConversationTurnObserverHandle>,
-    provider_round: usize,
-    followup_chain_active: bool,
-) -> ProviderTurnContinuePhase {
-    let tool_intents = turn.tool_intents.len();
-    let lane = preparation.lane_plan.decision.lane;
-    if tool_intents > 0 {
-        let running_tools_event =
-            ConversationTurnPhaseEvent::running_tools(provider_round, lane, tool_intents);
-        observe_turn_phase(observer, running_tools_event);
-        observe_provider_turn_tool_batch_started(observer, &turn);
-    }
-    let lane_execution = execute_provider_turn_lane(
-        config,
-        runtime,
-        session_id,
-        preparation,
-        &turn,
-        binding,
-        ingress,
-        observer,
-        followup_chain_active,
-    )
-    .await;
-    let should_emit_binding_trust_event =
-        !matches!(lane, ExecutionLane::Safe) || config.conversation.safe_lane_emit_runtime_events;
-    if should_emit_binding_trust_event {
-        emit_runtime_binding_trust_event_if_needed(
-            runtime,
-            session_id,
-            &lane_execution.turn_result,
-            binding,
-        )
-        .await;
-    }
-    observe_provider_turn_tool_batch_terminal(observer, &lane_execution.tool_events);
-    let loop_verdict = turn_loop_state.observe_turn(turn_loop_policy, &turn);
-    let followup_config =
-        ConversationTurnCoordinator::reload_followup_provider_config_after_tool_turn(config, &turn);
-    ProviderTurnContinuePhase::new(
-        tool_intents,
-        lane_execution,
-        loop_verdict,
-        followup_config,
-        ingress,
-    )
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
